requirements.txt
python-dotenv
requests
aiogram
redis
